add Tutorials refactor, UI improvements at settings.
