+++ client/src/components/pages/Store/Settings/Tutorials/index.js
@@ -1,0 +1,1 @@
+export * from "./Tutorials";
